Allow one form to navigate multiple frames consecutively
This patch allows one form to navigate two different frames by moving
all state associated with the form submission which occurs after the
async step as described in [1] from the form element and the
form element's document to the target frame being navigated.
Despite the spec saying that the task should be stored in the form
element, firefox and safari are both able to navigate multiple frames
like this. I plan to follow this up with a spec issue.

This organization of state reflect what we used to have with
NavigationScheduler [2], which was a member of LocalFrame.

I decided to use Frame instead of LocalFrame so that we can have the
same semantics for form submission regardless of whether we are
submitting to a LocalFrame or a RemoteFrame. The effects are observable
in this wpt: [3]

[1] http://html.spec.whatwg.org/C/#plan-to-navigate
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/html/forms/html_form_element.cc;l=519-520;drc=8f4e07de2e8797821a9446eb66ceeb305d763cbf
[3] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/external/wpt/html/semantics/forms/form-submission-0/form-double-submit-to-different-origin-frame.html

Bug: 1085097, 1047489
Change-Id: I1a212e52ef95449c5a0527d5530efaa49d384037